NoGUI: Split up the X11 and headless platforms from Main
